[SIGNALS] Added external signal support
This patch add a simple external interrupt handler.
We consider only bit changes on PORTC[2..7]. The interrupt handler
provides support to mark with a flash the "event" corresponding to the
associated PIN, than it is up to the control loop code to check
periodically for pending events and to take the proper actions.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>