[Control] Improved FAULTY and CALIBRATION timings
This patch improve the time required for FAULTY and CALIBRATION channels
identification. The approach is based on this channel selection logic:
1. channel status defines the priority to monitor a channel
2. an active higher priority channel is not switched until its
   priority do not decrease

Priorities are, from higher to lower: FAULTY, UNCALIBRATED and MONITORED.
A FAULTY channel deserve always the maximum priority among active ones.
If not faulty channels are active, but an UNCALIBRATED it is, this channel
preempt all other. Finally, on normal conditions, we cycle on active
channels that are on MONITORED state.

Moreover, on the first two state (i.e. FAULTY and UNCALIBRATED) we avoid
ADE7753 switching as much as possible, by checking if the current channel
is also active. In this case, we go ahead with it to faster identify
fault detection or calibration.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>