[Control] Added members to set/get enabled CHs
This patch make use of two control.h provided routine to access and manage
the bitmask of ENABLED channels.
Enabling/Disabling CHs has impact also on channels under calibration.
Thus, we export also the bitmask of calibration channel so that it could be
updated once the set of enabled channel is updated.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>