[LEDS] Improved GSM LEDs
The GSM link leds are initially all OFF. They are switched on, according
to the signal level, only when the modem is correctly initialized and we
are able to read the CSQ value.

Thus, if the modem is ON, and the signal LEDs are all OFF, there could
be a modem communication problem.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>