[GSM] Add one-time configuration support.
This patch provides support for one-time configuration of the GSM module.
An EEPROM flag is used to keep track of a configured flag, once the module
has been configured this flag is set. The next time the modem is powered
(or reset) the flag is checkf, if configured Autobounding and Configuration
is avoided.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>