[GSM] Bugfixed GSM restart
Once the GSM modem is restarted, it could produce a lot of output which
overflows the input FIFOs thus marking an error FLAG.
In this case, the read of the next AT command returns immediately because
of the pending error condition.

This patch, as well as flushing the FIFOs, clear all the error conditions
just before sending a new AT command.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>