[Control] Support for FAULT channel double checking
To try avoiding false alarms, this patch introduces a dummy double check
heuristic based on time. To assert a channel is fault it has to be measured
faulty for TWO times. Moreover, these two measurements are time separated
by a configurable time interval. Once a channel is detected fault the first
time it is "suspended" and no more checked for the next "pre-configured" time
interval [s]. After this time interval is elapsed, the channel is re-enabled
and could be still monitored. If the channel result to be FAULTY again this
second time, then a notification as usual is performed.

By default the channel suspension is 60[s]. However, this is one of the
configuration parameters which are expected to be exported for remote
configuration (e.g. by SMS).

Signed-off-by: Patrick Bellasi <derkling@gmail.com>